update sqla import
